+++ Jabberpoint_Sourcecode_Students/src/jabberPoint/interfaces/LoadAble.java
@@ -1,0 +1,10 @@
+package jabberPoint.interfaces;
+
+import jabberPoint.Presentation;
+
+import java.io.IOException;
+
+public interface LoadAble {
+
+    void loadFile(Presentation presentation, String fileName) throws IOException;
+}