+++ Jabberpoint_Sourcecode_Students/src/jabberPoint/interfaces/SaveAble.java
@@ -1,0 +1,10 @@
+package jabberPoint.interfaces;
+
+import jabberPoint.Presentation;
+
+import java.io.IOException;
+
+public interface SaveAble {
+
+    void saveFile(Presentation presentation, String fileName) throws IOException;
+}